post_filter: Add SuperRes multi-threading
Simply split the available rows by number of threads (rounding up
the left over rows to the main thread).

This requires the superres line buffer to be of size num_threads *
num_columns instead of simply num_columns.

On clips with superres on:
Performance on x86 is better by 15-45% depending on the clip/thread
count.
Performance on android (pixel 3) is better by 13-30% depending on
the clip/thread count.

PiperOrigin-RevId: 288558967
Change-Id: Ifc2431aaaf976a81b9aa138899fd1cf60510f969