paygen: delta_generator runs inside squawking thread
Steps that don't produce regular output get killed. We need
the delta_generator to do so, but since it's highly parallel,
we might not be outputting on time. This commit makes it so
it spawns a thread and squawks every minute to placate upstream.

BUG=chromium:1035799
TEST=./run_tests

Change-Id: Idb5a3c6dc113e23e01ad0174276be0361024692c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1990487
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>