[MR extension] Lower the severity of provideSinks(): provider not found error
When a native MRP is enabled instead of an extension one (DIAL or Cast), provideSinks() fails to find the extension MRP to provide sinks to, and logs an error. This behavior is expected until we stop providing sinks when not needed on the Chromium side, so the log's severity should be lowered.

Bug: crbug.com/1039386
PiperOrigin-RevId: 288607729