ShiftBuffer in opposite direction before decoding.
Call YuvBuffer::ShiftBuffer() in the opposite direction if the CDEF or
loop restoration process will call YuvBuffer::ShiftBuffer(). When we
return a decoded frame to the client, its plane buffer pointers should
be restored to the original position. This property will be important
when our frame buffer callbacks allow a client to allocate a frame
buffer (as opposed to a memory buffer) directly, because we won't be
able to shift the client's internal frame buffers.

PiperOrigin-RevId: 288750334
Change-Id: I774d7ede9395c56087f4a274141abd810693301a