Require Windows Vista / Server 2008 or greater.
This allows us to define pthread_cond_t as CONDITION_VARIABLE on
Windows.

I also expanded the only invocation of the WaitForSingleObject macro
and deleted the macro.

NOTE: I believe we can simply call the Windows WaitForSingleObject
function instead of WaitForSingleObjectEx(..., FALSE). The MSDN page for
WaitForSingleObject says:
  To enter an alertable wait state, use the WaitForSingleObjectEx
  function.
which implies that WaitForSingleObject enters a non-alertable wait state
and therefore WaitForSingleObject is equivalent to passing FALSE as the
bAlertable argument to WaitForSingleObjectEx.

Change-Id: Ief5b14b75cd3c8b02d2a26a67b3b2e6d8fa28ab2