Update en and fa dictionaries
This commit updates the en and fa dictionaries to their latest
upstream versions. It keeps the older dictionary versions so that
tests and Chrome Canary do not break. The older versions will be
deleted in a future commit. It also updates the download script
slightly to make the process of downloading the fa dictionary
smooth.

Change-Id: Ieb216f4e288982494d73fa341d57ff28f3d4da41