Pass frame buffer callbacks to BufferPool ctor.
Pass the frame buffer callbacks instead of the DecoderSettings struct to
the BufferPool constructor.

The motivation for this change is to allow us to pass the proposed new
frame buffer callbacks to the BufferPool constructor before exposing them
in the DecoderSettings struct. This allows libgav1 to test the new frame
buffer callbacks internally first.

PiperOrigin-RevId: 289118601
Change-Id: Iff6190b1ff60ec2502999ba7857be912da2742fc