Delete an out-of-date comment about ShiftBuffer.
The comment describes under what condition PostFilter shifts the plane
buffer pointers of a YuvBuffer. It mentions kFrameBufferRowAlignment,
which no longer exists, and does not describe the new ShiftBuffer calls
in PostFilter. Since it is not easy to keep the comment up to date, I
decided to just delete it.

PiperOrigin-RevId: 289131250
Change-Id: I1f1167c92b508d9947fb726ac31af3720dc10413