post_filter: Remove job queue for deblocking
For multi-threaded deblocking, we have a job queue which has the
row4x4, row_unit and plane. All those can be computed implicitly
from the incrementing job_number.

To keep the cache locality of the planes, we run the while loop in
each worker to jobs_per_plane * num_planes iterations. It will
process the Y plane for the first |jobs_per_plane| values of
|job_index|, and U plane for the next |jobs_per_plane| values of
|job_index| and so on.

This lets us avoid the allocation in this function for the job
queue and all the functions in the path can be turned to returning
void instead of bool.

PiperOrigin-RevId: 289183032
Change-Id: Ie0684ca29c80653ad42301ed5d4f300af132e0f1