Drop gandof and veyron_tiger from android-nyc-pfq.
They are going to covert to P in a very near future.
Before that happens, I want to make sure clearing all dependency
that is assuming the boards to be on N.

From now on after this CL, nyc-pfq is driven only by betty.
Once we are confident that we won't roll back to P, we can delete
the whole nyc-pfq at that point.

BUG=b:146630610
TEST=bot

Change-Id: Ic4d47632ca43e1231be8d635bec59759ef161c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/1999871
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Shao-Chuan Lee <shaochuan@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>