auto_updater.py: Be more gracious on AppID mismatches
If the device doesn't have an App ID or there is something wrong with
it, currenlty we assume cros flash is going to work. But actually, it
won't. This CL removes the condition (and generalizes the condition a
little bit better) so that we repair the properties file with empty App
ID for a partial match by nebraska.

BUG=b:147744816
TEST=cros flash

Change-Id: I1f4006d13e62bb5261d3059c9fee904ab7a00582
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2002679
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>