clean Tile::StoreMotionFieldMvsIntoCurrentFrame()
simpleperf shows CPU % dropped from 0.47% to 0.18% for 720P, 0.39% to
0.17% for 480P on an Android phone.

Only call this function when the current frame will be referenced.

Early terminate if the width or height to store is zero.

Store motion_field_reference_frame and motion_field_mv only when
necessary.

Add RefCountedBuffer::CleanMotionFieldReferenceFrame() to initialize
RefCountedBuffer::motion_field_reference_frame_, so that we can skip
some updates when the updates are the same as the initialized value.
It's fast and done in the frame level without multi-threading. I didn't
notice any CPU % increase for the caller function
DecoderImpl::DecodeTiles() after the new function is inlined.

PiperOrigin-RevId: 290080438
Change-Id: Ie03577b27d548b8b18768ba3e3f3014bbf1486f4