Add a C enum type Libgav1ImageFormat.
Make the C++ enum type ImageFormat an alias of Libgav1ImageFormat in the
libgav1 namespace. The original ImageFormat enumarators are defined as
constexpr constants, as suggested in this Stack Overflow article:
https://stackoverflow.com/questions/24212921/is-it-possible-to-alias-an-enum-class-enumerator

Change the version from 0.3.1 to 0.4.0 because the size of the
ImageFormat type changes from one byte (sizeof(uint8_t)) to 4 bytes.

Replace the bool is_monochrome, int8_t subsampling_x, and
int8_t subsampling_y parameters of the new frame buffer callbacks with
Libgav1ImageFormat image_format. This not only reduces the number of
function parameters but also avoids the use of the C++ 'bool' type in
these C functions.

PiperOrigin-RevId: 290351315
Change-Id: I8a1368fb7857d4926701d63ed744116b5c8b2058