Refactor Blend Chroma to compute one plane.
This is to prepare for multithreading. It makes more sense to have one
thread focus on one plane. Therefore, the DSP table entry is changed so
that it's called once for each plane.

The logic for whether to call CopyImagePlane is moved into AddNoise.
Thus the function is removed from film_grain_neon.cc. It also belongs in
an anonymous namespace, so it was moved from within film_grain.cc as
well.

PiperOrigin-RevId: 290828373
Change-Id: I34a737c0618c01516ba1264f487487fed2d9d846