cros_chrome_sdk: Don't disable the 'unknown-warning' clang warning.
This already gets set via the ebuild:
https://cs.corp.google.com/chromeos_public/src/third_party/chromiumos-overlay/chromeos-base/chromeos-chrome/chromeos-chrome-9999.ebuild?rcl=27cd2465ec90023bd52ba736de69343b34d27d8b&l=818

So we're adding the arg twice to cc and cxx commands. However, the
ebuild doesn't add that flag to the linker, while cros_chrome_sdk does.
So this is a no-op for cc and cxx commands, but removes the arg from
the linker when building via simple chrome.

Built simple chrome locally for arm-generic and amd64-generic and the
linker didn't complain, so should be safe.

BUG=chromium:937821
TEST=unittest & build chrome via simplechrome
Change-Id: I271fdb07bfd06878dd7ba8cfad279bd965ff135c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2020464
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>