warp: simplify region 2
Offsets are only important for the horizontal pass. When it is
skipped as in this case we can remove them entirely.

Continues the effort to consolidate the offset location for later
removal.

This may have an impact on the decision to separate the arm64/arm32
code but it's not worth a deeper investigation until the offset is
removed entirely.

PiperOrigin-RevId: 291452154
Change-Id: I3c2fff9a19b2c81afb84b6c42f496fa8939f0e72