M80 merge: Preserve scissor for RPDQs with filters
Cherry pick of https://chromium-review.googlesource.com/c/chromium/src/+/2019804

If the RPDQ has a filter, it's touched pixels are not actually restricted
to the visible rect of the quad. In that case it is incorrect to explicitly
clip the visible rect to the scissor and not set the scissor as a clipRect.
This CL makes it so the scissor is remembered and is applied post-filtering,
so effects like drop shadows are properly clipped to the window content.

Bug: 1035271
Change-Id: I138b1412c55489aa0068cc0ea1744a3248738716
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023350
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/branch-heads/3987@{#721}
Cr-Branched-From: f8ec429d946866f630bf422440e162a16c8b97d1 [formerly c4e8da9871cc266be74481e212f3a5252972509d]-refs/heads/master@{#722274}

Former-commit-id: 5cf10da1527d2cb4bc3bb7a1d3d143f44a1d8748