cros_chrome_sdk: Don't add the toolchain dir to the compiler's cmd line.
The code being removed here was added so that the compiler could find
the right gold linker. However, we're no longer using gold now that
we've switched to clang + lld. And since clang can find the lld that
sits right next to it, we shouldn't need to add the toolchain dir
to compiler invocations.

BUG=chromium:937821
TEST=built simplechrome locally
TEST=ran this on chrome's simplechrome bots: crrev.com/c/1863361/25
Change-Id: Idc65a2d3a5f183d1a7b75a2535d062cfad567ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2029111
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>