Get buffer private data if using callback adaptor.
Get the real buffer private data from the version 1 FrameBuffer struct
when the frame buffer callback adaptor is used. In that case,
buffer_private_data_ points to a version 1 FrameBuffer struct and is not
the real buffer private data.

This bug shows the emulation provided by the frame buffer callback
adaptor is not completely transparent.

This bug is reported by Andrew Lewis:
https://github.com/google/ExoPlayer/issues/6928

PiperOrigin-RevId: 292384088
Change-Id: Iebe6e44c9df0ee119fd3c5b1a2a9ec2e2d60a233