mDNS: Add support for publishing multiple PTR records
Previously, only publishing one PTR record at a time was supported,
because we did not have a use case for publishing multiple. In order to
allow for more extensive E2E tests, it is helpful to support publication
of more than one. This CL adds that support.

Change-Id: Icbe91ccef47ce54ea954efc8693ad5c5759ebbbe
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2031687
Commit-Queue: Ryan Keane <rwkeane@google.com>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>