Fix GCC + libsdl/libavcodec build
Currently, if FFMPEG and SDL support is enabled and the compiler is set
to GCC, we get a build error on some versions due to a macro  that
results in a compile error. This patch removes the macro usage in favor
of using a properly instantiated buffer.

Change-Id: I5f95ab268296306e4808ce621bfb356b3f141dd7
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2036490
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>