reduce quantized coefficient padding
Removed quantized coefficient column padding, and reduced row padding for 2D and
horizontal.
These saves the memory size to be initialized. Furthermore, PaddedIndex() is removed.
Also it's easy to make quantized_buffer[] 16-byte aligned for each row if necessary.
Although we added a boundary check in ReadCoeffBase{Horizontal, Vertical}, this check is
fast and these two functions are not frequently called.

Decoder is ~0.5% faster on an Android phone.

PiperOrigin-RevId: 292963066
Change-Id: I5c1a094feb616819f6bfe225752fea99c2833914