warp: rename is_clip to is_compound
This is not a necessary change (is_clip is used to distinguish between the
warp functions) but it is a lot clearer and matches the convolve distinction.

With |inter_round_bits_vertical| derived from the function being called we
can move forward with removing it from the warp and convolve parameters.

PiperOrigin-RevId: 292978880
Change-Id: I318b6215e10608d04bd1f718c0dd6cbe1c7e9d69