Add a C API.
1. Use the current public header files for the C API. Do not add new
public header files.

Every public header has a C section at the top and a C++ section (guarded
by the __cplusplus macro) at the bottom.

2. The enums and structs are defined as C types with the "Libgav1"
prefix. The original C++ enum and struct names become type aliases in
the libgav1 namespace.

C++ enumerators (the values of an enum type) become constexpr constants
in the libgav1 namespace. This is the closest approximation I found and
seems to work well.

There are two limitations:

* In C++, we can no longer specify underlying type (e.g., uint8_t) for
  an enum type.

* The structs cannot have constructors. Only one struct,
  DecoderSettings, has a constructor. The constructor is replaced by the
  Libgav1DecoderSettingsInitDefault() function.

Note: Unlike the other enums and structs, DecoderSettings cannot be
defined as a type alias for Libgav1DecoderSettings for two reasons:
* DecoderSettings has a bool data member. The bool type is not available
  in C.
* DecoderSettings has a constructor.

3. Each C++ function libgav1::DoSomething() is converted to the C
function Libgav1DoSomething(). The original C++ function is defined as
an inline function that calls the new C function.

4. Each C++ method Foo::DoSomething() has a corresponding C function
Libgav1FooDoSomething() that takes a pointer to the Foo object as the
first argument (if the method is non-static) and internally calls the
Foo::DoSomething() method.

PiperOrigin-RevId: 293013954
Change-Id: I961a16c1c296bc65953d84f35421ab4e4a3df8f0