system_api: vm_concierge: add option to generate digest when exporting a VM
For automatic deployment of Plugin VMs we expect administrators to
create a policy with a link to a VM image and SHA256 digest of that
image, so that we can verify that downloaded image is not damaged or
altered mid-flight. This change extends the definition of "export VM"
concierge command to allow requesting said SHA256 digest when creating
an image of existing VM.

The digest is expected to be written into the 2nd file descriptor
attached to the DBus message. We are using a file descriptor instead of
returning digest in a reply or signal because export operation takes
long time and CLI may elect not to wait for it to complete before
returning control to the user and having it in a file means that it can
be examined later.

BUG=b:148281196
TEST="vmc export -d TestVm testvm.zip", ensure that digest matches
output of "sha256sum ~/Downloads/testvm.zip"

Change-Id: Idb856cefe6eef2f310612df82c782a04b4e3935f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2039936
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: f5534aa450d7ae1ac80af73a9fb9ee90072947e7