Cast Streaming Sender: Introduce SenderPacketRouter.
This is the entity that manages network packet transmission (inbound and
outbound) for one or more Senders sharing the same UDP socket. It routes
each inbound packet to a specific Sender instance. For outbound packets,
it paces packet transmission, scheduling burst-sends, and respecting the
priority of RTCP-before-RTP and one Sender over the other.

Bug: openscreen:55
Change-Id: If3d9faefeb634188906212183d6306429575e9e5
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2038462
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Commit-Queue: Yuri Wiitala <miu@chromium.org>