run_tests: switch tool & unittest to be Python 3 only
We add a bit of logic to run_tests so that it doesn't bother testing
against python2 since it's using the new wrapper3.py.  Then we cut
both run_tests and its unittests over to be Python 3 only.  This is
safe to do as nothing imports run_tests, only executes it.

BUG=chromium:997354
TEST=`./run_tests` passes

Change-Id: I845e982b155e9df5375277313ddb3e466ee6b53e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2040300
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>