fwgdb: Use ServodClientParser
The script was throwing an exception because opts.serialname was not
specified. I just switched over to using the ServodClientParser so all
the correct args are defined. This does mean that -s had to be changed
to -y because -s is used for --serialnumber.

BUG=none
TEST=fwgdb -b zork -c -r no
Was able to connect to the dut.

Change-Id: I33358c57714a630472e5626600817fa8b0809b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2041692
Auto-Submit: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Mathew King <mathewk@chromium.org>