lib/git: break out ChangeId into its own function
Code in Commit() will get the Change-Id after doing a git commit,
but in some use cases, we want to get the Change-Id of a commit
that already exists, without changing it. Move this code into its
own function that Commit() calls or an external user can call it
directly.

BUG=None
TEST=Run unit test, also check it manually
$ ./git_unittest
........................................
----------------------------------------------------------------------
Ran 40 tests in 8.031s

OK
$ python3
>>> from chromite.lib import git
>>> print(git.GetChangeId('.'))
>>> exit(0)
$ git show
Compare the Change-Id printed by the python code with that shown

Change-Id: Ia7b712c42ff83c52c0fb5d88d1ef6c62f49da88d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2041804
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>