gerrit: rename label utility funcs
We're running into the problem where our colloquial names for labels
are colliding with Gerrit endpoints/concepts.  Gerrit now has the
concept of "ready" to complement "work-in-progress", but we've been
using "ready" to mean "set the Commit-Queue label".

Lets rename our helpers that exist only to set labels to have a better
scoped name.  Hopefully this will avoid any future collisions, while
making it clearer what's going on, and be more portable across other
Gerrit instances.

There will be a PSA for the community for the breakage, and we'll hold
off adding the new "ready" and "review" helpers for a bit.

BUG=None
TEST=`gerrit -n label-as 1234 4`

Change-Id: I12429b3c6108b2561273d72dcca504357579ccf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2042040
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>