tile: Store motion field MVs in the parse step
When split_parse_and_decode_ is true, we currently store motion
field mvs into the current_frame after the decode step. It can
actually be done as soon as the parsing is complete. It paves the way
for frame parallel by minimizing the parsing dependency between
frames (it also keeps the parse thread a bit more busy).

PiperOrigin-RevId: 293722325
Change-Id: Iff4d04c04d8f260f52fde3660655c50c1dbdb91e