gerrit: rewrite helpers to use classes
This adds more LOC, but it makes the parsing more robust, and allows
for better encapsulation of subcommand-specific options.  The current
code forces all options for all subcommands to be part of the single
global parser leading to a lot of options that only make sense for a
few subcommands.  It also makes it hard to build more complicated
subcommands with multiple optional settings.

BUG=None
TEST=ran various `gerrit` commands

Change-Id: Ice2bb390f6c542325886e8b2da73c4b426ea9803
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2048210
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>