Add hybi.compute_accept_from_unicode()
The current implementation of compute_accept only allows bytes strings
as its argument. However, when computing accept keys from handshake
headers, there is a need to encode unicode strings to bytes strings and
pass it on to compute_accept. This commit provides
compute_accept_from_unicode which does the encoding before passing it on
the compute_accept so that websocket handlers do not need to encode
everytime.

This commit also modifies compute_accept by removing binary_accept from
return which was only used in logging.

Change-Id: Ideb50bbfdc25bc85de53be8fea4f0fd3a2834683