Use a dynamic number of frame buffers
Storing a RefCountedBuffer object in a Vector seems to be
complicated because of the copy/move semantics. So the simplest
way around that is to store a list of pointers in the vector.

The number of frame buffers need to be dynamic because in the
frame parallel case we could be allocating more than kNumBuffers
frames at a given time since we could be decoding multiple frames
at the same time.  To avoid the max calculation, it is simpler to
merely use a dynamic array. This will not change the behavior much
in the non frame parallel case.

PiperOrigin-RevId: 293890169
Change-Id: Ib2b7b5c81b67c9e27c298c5ee8e58291ff1f3a17