[Open Screen] Allow PlatformClientPosix to take a TaskRunner.
This updates the API to allow clients to pass a TaskRunner, so they
don't have to subclass PCP.

It also streamlines the documentation, fixes IWYU, and reorders
definitions to match declaration order.

Finally it also declares TaskRunner::IsRunningOnTaskRunner as pure
virtual so it is mandatory to implement.

Bug: openscreen:108
Change-Id: I646f416d7d337f32c96186236a84a9b9440f1eeb
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2043953
Commit-Queue: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Ryan Keane <rwkeane@google.com>
Reviewed-by: Jordan Bayles <jophba@chromium.org>