Crash fix: Prevent a unit test from using a real UDP socket.
The code in receiver_session_unittest.cc was creating an Environment
instance that was binding a real UDP socket. Also, when the test was
running in a Chrome embedder build of the openscreen_unittests, the
result was a segfault because a full browser environment was not
present.

To fix this, a MockEnvironment class (subclass of Environment) is used
to intercept socket calls and provide a fake local address/port for the
tests to work with.

Change-Id: I59a06ed12b58a1ba14d228cad7cd8dfffafc9bb2
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2051319
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Yuri Wiitala <miu@chromium.org>