mDNS: NSEC support in Querier
This CL adds support for negative response NSEC records to the
MdnsQuerier class. NSEC records are used to signify that a given
record does NOT exist, as detailed in RFC 6762 section 6.1.

Specifically, this CL updates the MdnsRecordTracker such that it can
store either a record of a given type and class or an NSEC record which
shows that record's nonexistence. Further changes are made to support
this change throughout MdnsQuerier.

Change-Id: I6cdf1a48035e8a2760751870530ff8f9881eca00
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2039343
Commit-Queue: Ryan Keane <rwkeane@google.com>
Reviewed-by: Jordan Bayles <jophba@chromium.org>