prediction: remove offsets for bitdepth == 8
Convolve calculations use reduced multipliers which removes
the need for an offset while calculating.

Warp still requires the offset for at least the first pass.
With some investigation it should be possible to simplify
things a little bit further.

Removing the offset from the warp calculation before storing
it saves effort in the later blending process.

PiperOrigin-RevId: 295208440
Change-Id: I3c06cab829d407c165d9028986654724b3d87d50