tile: Generalize DecodeSuperBlockRow
For the frame parallel mode, we will have to parse and decode each
tile in separate passes. Rename DecodeSuperBlockRow to
ProcessSuperBlockRow and make it a template which takes in the
processing mode (parse/decode/both) as a template parameter and
a boolean indicating whether or not we should save the symbol
decoder context state after processing the last row.

PiperOrigin-RevId: 295235094
Change-Id: I6507da854829d4f159aca1184ddb0f31d54c8237