service: packages: fix encoding with android uprev
This code runs a program & parses its output as text, but doesn't set
the encoding, so it was getting binary data back.  The tests didn't
catch this error because it mocked things out at the wrong level.

Switch the tests over to using proper command mocks which catches the
encoding failure, then fix the uprev code to use encoding=utf-8.  We
also get to collapse the _parse_android_atom function because the
whole reason for it existing (for unittest mocks) is no longer needed.

BUG=chromium:1052524
TEST=`./run_tests` passes

Change-Id: Id5bb43a355dfbf7b969802d6968305a71152c139
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2058044
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>