Remove aligned_width() and aligned_height().
We have been using the aligned width and height in the calculation of
frame buffer size. Use the real width and height instead. The border
sizes we use more than cover the 0 to 7 pixels of padding in the aligned
width and height.

In addition, pass the real width and height, instead of the aligned
width and height, to frame buffer callbacks. (This fixes a bug because
some callbacks need the real frame width and height.) As a result, frame
buffer callbacks can no longer assume the width and height parameters
are even numbers and will need to take care when subsampling the width
and height for the chroma planes.

Note: It seems that aligned_width() and aligned_height() are equivalent
to MiCols * MI_SIZE and MiRows * MI_SIZE in the spec. I copied
aligned_width() and aligned_height() from libvpx and libaom. In libgav1
we calculate those values directly. All prior uses of aligned_width()
and aligned_height() have been found to be incorrect and removed.

PiperOrigin-RevId: 295552312
Change-Id: I0221eb673844a46a0bce4a983de6c355cab4cfae