Cast pointer |addr| to uintptr_t in aom_align_addr
The aom_align_addr() macro needs to cast the pointer parameter |addr|
to an integer. In C99 we can cast the pointer to uintptr_t, a type
designed specifically to be capable of holding a pointer.

Note: The uintptr_t type is already used in av1/encoder/hash.c and
av1/av1_dx_iface.c, so it is safe to use uintptr_t in libaom.

Change-Id: I63b6e602920676380e4188d21af00edd525a3d38