export_to_cloud_trace: get it working under Python 3
We need to rework the generator to account for PEP-479 changes where
nested StopIteration exceptions are changed to RuntimeError.

The _GroupBy logic needs a tweak in case the entries are missing the
traceId field as Python 3 throws errors if you try to compare None
using > or < operators.

BUG=chromium:997354
TEST=CQ passes

Change-Id: Ida1ca9450e6adb6fd9a00e18941312891667670d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2063222
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>