cbuildbot: use goma for latest_toolchain builds
The toolchain team would quite like to be able to see all of the cool
`.ninja_log` visualizations/etc that we can get with
non-latest_toolchain builds.

Unfortunately, the bits that upload that seem pretty tightly coupled
with all of goma's other logging.

This CL does everything but turning `USE_GOMA` on for latest-toolchain
builds. It's a hacky approach, but given that cbuildbot is deprecated,
and that this hack is relatively small, I'm not sure that's the end of
the world.

Actually setting `USE_GOMA=1` would be quite painful, since AIUI local
fallbacks use 15-16 cores, which would substantially increase Chromium
build times.

BUG=chromium:1054897
TEST=tryjobs
Change-Id: I32ff635bbeb88fe20e1ffad20c23f48ebcaae971
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2067503
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@google.com>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>