Don't consider extra ID3 tags as a fatal error
If two copies of ID3 tags are found, ffmpeg treats it as a fatal error
if |error_recognition| is set to a high level. However, it makes sense
to simply discard the second one, as some files are encoded with both
(probably for compatibility with different players). Commenting out the
failure to avoid problems playing these files.

Was removed in the past, looks like the recent merge added it back.
This is basically redoing 7262141e8c14462f47e79ed6ba457321dea9cc5d.

Bug: 891179,1051234
Test: external/wpt/webaudio/the-mediaelementaudiosourcenode-interface/ tests pass
Change-Id: Id58887a534ede687d9cfee40ac9116ba48cf4250
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/2067997
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>