Remove declaration of InitializeScalingLookupTable
The InitializeScalingLookupTable method is not defined, so we should not
declare it.

This issue is detected by a Visual C++ warning. I fixed this issue in
CL 291933222. It came back in CL 293253364 because src/film_grain.h
branched from the revision of src/dsp/film_grain_impl.h immediately
before my fix.

PiperOrigin-RevId: 296551579
Change-Id: Ide6b571ec7bf1ae0deb95d434a46943d9793f567