afdo: Add verifiers to publish 4.14 and 4.19 kernel profiles.
As a part of rolling out kernel AFDO profiles for 4.14 and 4.19,
we plan to enable them on one board from each kernel version. See
crrev.com/c/2064544. This patch adds support to add verification
builders to automatically update the profile to the latest. It
only impacts the boards that enables the profiles (two for now).

BUG=chromium:1026315
TEST=None

Change-Id: I4fe8b4ad441cd763781cc05bcdf5aba6b37a4fb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2070556
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>