afdo: Stop updating kernel 3.14 AFDO profiles in legacy PFQ.
In legacy PFQ (still used on branches), we used to set a deadline
for AFDO profiles to prevent using very profiles. Since we stopped
to update 3.14 profiles on master, we also need to make change to
legacy PFQ, which is still being used on branches.

BUG=chromium:1055452
TEST=None

Change-Id: Ie0b413e01c101388b44076e2572c1cd25cf72b57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2070633
Commit-Queue: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>