upload_symbols: Update call to pass in file data versus file handler
We are currently experiencing an issue in which a single symbol file
fails for all boards.  This is a repeatable error but only happens with
the one file, which doesnt appear to show any different characteristics.
Updating the call to upload to take in the read file data, versus the
actual file handler, appears to resolve this issue.

BUG=chromium:1055332
TEST='cros_sdk -- /mnt/host/source/chromite/run_tests'

Change-Id: I2a6705550e6965dd90fd0b2e01782ec7545292b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2071492
Tested-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>