vm_tools system_api: Fix the name of a field.
disk_path is actually vm_name. It's treated as the name of the VM by
code, so this doesn't change the meaning of the field, and there are no
callers yet so it's safe to rename.
This is incorrect elsewhere in Concierge as well, but those are in use
so fixing them is a much larger project to do separately.

BUG=chromium:858815
TEST=CQ

Change-Id: Id211c5f783de4d44369510d762469932223000da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2072977
Commit-Queue: David Munro <davidmunro@google.com>
Tested-by: David Munro <davidmunro@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Auto-Submit: David Munro <davidmunro@google.com>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 7820dab49ea6003450b13bf0f74f3e263b5ef2ba