cbuildbot: Add 'no-publish-prebuilts-conf' flag
This approach would enable running a production instance of the SDK
builder based on a branch and having that builder upload binary packages
but *not* write any changes to prebuilt.conf or sdk_version.conf. A
person could then manually update the prebuilt.conf paths on the branch
to point to these uploaded artifacts as a way of providing new
cross-compilers on a release branch.

BUG=chromium:1055914
TEST=None

Change-Id: I622af5b2d8a41e7afcb02c86b22a9f07f3b11e6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2072942
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>