Cast Streaming: Add GetRecentReportTime() to SenderReportBuilder.
During code review of the Sender, it was noticed that there is no need
to maintain a cache that maps recent StatusReportIds to

Clock::time_point values (that represent when a Sender Report was sent).
This change adds a simple utility method to compute the values directly
from the StatusReportIds.
Change-Id: I340ecc5473447bf60e67d4e4bf75457e638035f7
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2077159
Reviewed-by: Max Yakimakha <yakimakha@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Commit-Queue: Yuri Wiitala <miu@chromium.org>