Restrict python-spawned processes to the initiator (#706)
This commit disables Incredibuild's CreateProcess interception on
python.  This effectively forces all child processes spawned from python
to the same machine.  Since python is already restricted to the initiator
machine, this will make all its child processes to also be restricted to
the initiator machine.

Without this restriction, the build fails with error 22 when the `brotli.exe`
executable is attempted to be launched. Restricting `brotli` alone
doesn't seem to help. Even if that was possible, `brotli` incurs a lot
of I/O with its parent `python` process so having them on separate
machines would incur a lot of network I/O. This may not have any
performance benefit or may even hurt performance due to high network
I/O.