auto_updater: update CheckPayloads method to ssh into devserver
When a devserver is in a restricted subnet, the CheckPayloads() method
in LabTransfer fails as it tries to run a curl command directly instead
of sshing into the devserver. This CL changes the method so that the
CheckPayloads() method SSHs into the devserver first.

BUG=chromium:1054848
TEST=Manual tests, unit tests

Change-Id: I46e3172af760584e20e0d09714065647e1b03fd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2086471
Commit-Queue: Sanika Kulkarni <sanikak@chromium.org>
Tested-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Auto-Submit: Sanika Kulkarni <sanikak@chromium.org>