chromeos_config: Introduce a new loon builder tied to release tool
Loon is utilizing ChromeOS build infrastructure but are being hindered
by a lack of resources in their associated build pool; they are not
using release artifacts, and none are being uploaded, so they instead
utilize tryjobs which are a shared resource.  This will associate a new
type of builder, based on release configuration, that is tied to a
different release tool.  All new builders must end with -rapid and will
bypass the template check.

BUG=chromium:1056001
TEST='cros_sdk -- /mnt/host/source/chromite/run_tests'

Change-Id: Ib76671ef582ff433cd0a9f36ae5617e681b3eda9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2078960
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>