Eliminate YuvBuffer::ShiftBuffer
Rework the PostFilter and Tile co-ordination to eliminate the
YuvBuffer::ShiftBuffer function.

We still have a logical shift but it is co-ordinated by the
PostFilter and Tile classes without actually shifting the
YuvBuffer. The Tile class gets the plane pointer to write the
decoded output to from the PostFilter class.

PiperOrigin-RevId: 297902031
Change-Id: Ic181fef5c70fac83e59a873d6b6317a0c22923b0