cros build-ap: Add FW_NAME support.
Add support for the FW_NAME envvar support. This allows only a specific
variant to be built, saving time when iterating on a specific variant's
firmware.

Also remove the --board argument option so it falls in line with the
long term standards, and simplify and harden the build target argument
parsing.

BUG=chromium:1057201
TEST=./run_tests
TEST=cros build-ap -b volteer  still works.

Change-Id: I0727fbf49e91a504c19ed0d5e3385ca1a5cd08f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2080646
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>