auto_updater: Fix error message.
If update_engine is in the UPDATE_STATUS_IDLE state after downloading
and finalizing an update, the auto updater tries to fetch the
update_engine log. This is done using RunCommand with a string
argument, but RunCommand only support list arguments. This change
corrects the call.

BUG=chromium:1057877
TEST=cros flash a DUT that is having problems updating.
  run_tests

Change-Id: I63cf464df7530d7a6f3aad22a22e5633d584bf13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2086473
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Tested-by: Laurence Goodby <lgoodby@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>